# code/clients/typescript/openapi/booking.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Booking API",
    "version": "v1"
  },
  "paths": {
...
    "/erasure/check-active": {
      "post": {
        "tags": [
          "Erasure"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceErasureInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceErasureInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceErasureInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/erasure": {
      "post": {
        "tags": [
          "Erasure"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceErasureInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceErasureInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceErasureInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServiceErasureInput": {
        "required": [
          "erasureRequestId",
          "tenantId",
          "targetActorHash"
        ],
        "type": "object",
        "properties": {
          "erasureRequestId": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "targetActorHash": {
            "type": "string"
          },
          "targetUserId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Bearer — tenant and user identity come from token claims, not request parameters.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}